+++ clients/fides-js/src/lib/i18n/i18n-utils.ts
@@ -5,6 +5,7 @@ import {
   FidesExperienceTranslationOverrides,
   FidesInitOptions,
   PrivacyExperience,
+  PrivacyExperienceMinimal,
   PrivacyNotice,
   PrivacyNoticeTranslation,
 } from "../consent-types";
@@ -437,7 +438,7 @@ export function selectBestNoticeTranslation(
  */
 export function selectBestExperienceConfigTranslation(
   i18n: I18n,
-  experience: ExperienceConfig,
+  experience: Partial<ExperienceConfig>,
 ): ExperienceConfigTranslation | null {
   // Defensive checks
   if (!experience || !experience.translations) {
@@ -650,7 +651,7 @@ export function setupI18n(): I18n {
 export const localizeModalLinkText = (
   disableLocalization: boolean,
   i18n: I18n,
-  effectiveExperience?: Partial<PrivacyExperience>,
+  effectiveExperience?: Partial<PrivacyExperience | PrivacyExperienceMinimal>,
 ): string => {
   let modalLinkText = DEFAULT_MODAL_LINK_LABEL;
   if (!disableLocalization) {